HiddenMarkovModels

Documentation for HiddenMarkovModels.

HiddenMarkovModels.HiddenMarkovModels

HiddenMarkovModels.AbstractHMM

HiddenMarkovModels.AbstractHiddenMarkovModel

Base.rand

HiddenMarkovModels.backward!

HiddenMarkovModels.baum_welch!

HiddenMarkovModels.emission_distribution

HiddenMarkovModels.emission_distributions

HiddenMarkovModels.fit_mle_from_multiple_sequences

HiddenMarkovModels.fit_mle_from_single_sequence

HiddenMarkovModels.forward!

HiddenMarkovModels.forward_backward!

HiddenMarkovModels.initial_distribution

HiddenMarkovModels.is_prob_vec

HiddenMarkovModels.is_square

HiddenMarkovModels.is_trans_mat

HiddenMarkovModels.marginals!

HiddenMarkovModels.nb_states

HiddenMarkovModels.rand_prob_vec

HiddenMarkovModels.rand_trans_mat

HiddenMarkovModels.transition_matrix

HiddenMarkovModels.HiddenMarkovModels — Module

HiddenMarkovModels

A Julia package for HMM modeling, simulation, inference and learning.

HiddenMarkovModels.AbstractHMM — Type

AbstractHMM

Shortcut for AbstractHiddenMarkovModel.

HiddenMarkovModels.AbstractHiddenMarkovModel — Type

AbstractHiddenMarkovModel

Interface for Hidden Markov Models with arbitrary emissions.

Required methods

nb_states(hmm, θ)

initial_distribution(hmm, θ)

transition_matrix(hmm, θ)

emission_distribution(hmm, i, θ)

Base.rand — Method

rand(rng, hmm, T, θ)
rand(hmm, T, θ)

Simulate hmm for T time steps and return a named tuple (state_seq, obs_seq).

HiddenMarkovModels.backward! — Method

backward!(β, bβ, A, B)

Perform a backward pass by mutating β and bβ (after forward pass).

HiddenMarkovModels.baum_welch! — Method

baum_welch(hmm::HMM, obs_seqs; max_iterations, tol)

Apply the Baum-Welch algorithm on multiple observation sequences, starting from an initial estimate hmm.

HiddenMarkovModels.emission_distribution — Function

emission_distribution(hmm, θ, i)

Return the emission distribution in state i for hmm with parameters θ.

The result must satisfy the DensityInterface.jl specification.

HiddenMarkovModels.emission_distributions — Method

emission_distributions(hmm, θ)

Return the vector of emission distributions for hmm with parameters θ.

Each element of the result result must satisfy the DensityInterface.jl specification.

HiddenMarkovModels.fit_mle_from_multiple_sequences — Method

fit_mle_from_multiple_sequences(::Type{D}, xs, ws)

Fit a distribution of type D based on multiple sequences of observations xs associated with multiple sequences of weights ws.

Must accept arbitrary iterables for xs and ws.

HiddenMarkovModels.fit_mle_from_single_sequence — Method

fit_mle_from_single_sequence(::Type{D}, x, w)

Fit a distribution of type D based on a single sequence of observations x associated with a single sequence of weights w.

Defaults to Distributions.fit_mle, with a special case for vectors of vectors (because Distributions.fit_mle accepts matrices instead). Users are free to override this default for concrete distributions.

HiddenMarkovModels.forward! — Method

forward!(α, c, p, A, B)

Perform a forward pass by mutating α and c.

HiddenMarkovModels.forward_backward! — Method

forward_backward!(α, c, β, bβ, γ, ξ, p, A, B)

Apply the full forward-backward algorithm by mutating α, c, β, bβ, γ and ξ.

HiddenMarkovModels.initial_distribution — Function

initial_distribution(hmm, θ)

Return the vector of initial state probabilities for hmm with parameters θ.

HiddenMarkovModels.is_prob_vec — Method

is_prob_vec(p; atol)

Check if p is a probability distribution vector.

HiddenMarkovModels.is_square — Method

is_square(A)

Check if A is a square matrix.

HiddenMarkovModels.is_trans_mat — Method

is_trans_mat(A; atol)

Check if A is a transition matrix.

HiddenMarkovModels.marginals! — Method

marginals!(γ, ξ, α, β, bβ, A)

Compute state and transition marginals by mutating γ and ξ (after backward pass).

HiddenMarkovModels.nb_states — Function

nb_states(hmm, θ)

Return the number of states of hmm with parameters θ.

HiddenMarkovModels.rand_prob_vec — Method

rand_prob_vec(rng, N)
rand_prob_vec(N)

Generate a random probability distribution vector of length N.

HiddenMarkovModels.rand_trans_mat — Method

rand_trans_mat(rng, N)
rand_trans_mat(N)

Generate a random transition matrix of size N × N.

HiddenMarkovModels.transition_matrix — Function

transition_matrix(hmm, θ)

Return the matrix of state transition probabilities for hmm with parameters θ.