API reference

HiddenMarkovModels — Module

HiddenMarkovModels

A Julia package for HMM modeling, simulation, inference and learning.

Types

HiddenMarkovModels.AbstractHiddenMarkovModel — Type

AbstractHiddenMarkovModel

Abstract supertype for an HMM amenable to simulation, inference and learning.

Interface

To create your own subtype of AbstractHiddenMarkovModel, you need to implement the following methods:

length(hmm)

eltype(hmm, obs)

initialization(hmm)

transition_matrix(hmm)

obs_distributions(hmm)

fit!(hmm, init_count, trans_count, obs_seq, state_marginals) (optional)

Applicable functions

Any HMM object which satisfies the interface can be given as input to the following functions:

logdensityof(hmm, obs_seq)

forward(hmm, obs_seq)

viterbi(hmm, obs_seq)

forward_backward(hmm, obs_seq)

rand(rng, hmm, T)

baum_welch(hmm, obs_seq) (if fit! is implemented)

HiddenMarkovModels.HiddenMarkovModel — Type

struct HiddenMarkovModel{I<:(AbstractVector), T<:(AbstractMatrix), D<:(AbstractVector)} <: AbstractHiddenMarkovModel

Basic implementation of an HMM.

Fields

init::AbstractVector: initial state probabilities

trans::AbstractMatrix: state transition matrix

dists::AbstractVector: observation distributions (must be amenable to logdensityof and rand)

HiddenMarkovModels.AbstractHMM — Type

AbstractHMM

Alias for the type AbstractHiddenMarkovModel.

HiddenMarkovModels.HMM — Type

HMM

Alias for the type HiddenMarkovModel.

Basics

Base.rand — Function

rand(hmm, T)
rand(rng, hmm, T)

Simulate hmm for T time steps.

Base.length — Function

length(hmm)

Return the number of states of hmm.

Base.eltype — Function

eltype(hmm, obs)

Return a type that can accommodate forward-backward computations on observations similar to obs.

It is typically a promotion between the element type of the initialization, the element type of the transition matrix, and the type of an observation logdensity evaluated at obs.

HiddenMarkovModels.initialization — Function

initialization(hmm)

Return the vector of initial state probabilities for hmm.

HiddenMarkovModels.transition_matrix — Function

transition_matrix(hmm)

Return the matrix of state transition probabilities for hmm.

HiddenMarkovModels.obs_distributions — Function

obs_distributions(hmm)

Return a vector of observation distributions, one for each state of hmm.

There objects should support rand(rng, dist) and DensityInterface.logdensityof(dist, obs).

Inference

DensityInterface.logdensityof — Function

logdensityof(hmm, obs_seq)

Run the forward algorithm to compute the posterior loglikelihood of sequence obs_seq for hmm.

This function returns a number.

HiddenMarkovModels.forward — Function

forward(hmm, obs_seq)

Run the forward algorithm to infer the current state of hmm after sequence obs_seq.

This function returns a tuple (α, logL) where

α[i] is the posterior probability of state i at the end of the sequence

logL is the loglikelihood of the sequence

HiddenMarkovModels.viterbi — Function

viterbi(hmm, obs_seq)

Apply the Viterbi algorithm to infer the most likely state sequence corresponding to obs_seq for hmm.

This function returns a vector of integers.

HiddenMarkovModels.forward_backward — Function

forward_backward(hmm, obs_seq)

Run the forward-backward algorithm to infer the posterior state and transition marginals of hmm on the sequence obs_seq.

This function returns a tuple (γ, ξ, logL) where

γ is a matrix containing the posterior state marginals γ[i,t]

ξ is a vector of matrices containing the posterior transition marginals ξ[t][i,j]

logL is the loglikelihood of the sequence

See also

ForwardBackwardStorage

HiddenMarkovModels.baum_welch — Function

baum_welch(hmm_init, obs_seq; kwargs...)
baum_welch(hmm_init, obs_seqs, nb_seqs; kwargs...)

Apply the Baum-Welch algorithm to estimate the parameters of an HMM starting from hmm_init.

Return a tuple (hmm_est, logL_evolution).

Keyword arguments

atol: minimum loglikelihood increase at an iteration of the algorithm (otherwise the algorithm is deemed to have converged)

max_iterations: maximum number of iterations of the algorithm

loglikelihood_increasing: whether to throw an error if the loglikelihood decreases

StatsAPI.fit! — Function

fit!(hmm, init_count, trans_count, obs_seq, state_marginals)

Update hmm in-place based on information generated during forward-backward.

This method is only necessary for the Baum-Welch algorithm.

Arguments

init_count::Vector: posterior initialization counts for each state (size N)

trans_count::AbstractMatrix: posterior transition counts for each state (size (N, N))

obs_seq::Vector: sequence of observation, possibly concatenated (size T)

state_marginals::Matrix: posterior probabilities of being in each state at each time, to be used as weights during maximum likelihood fitting of the observation distributions (size (N, T)).

See also

BaumWelchStorage

ForwardBackwardStorage

Misc

HiddenMarkovModels.rand_prob_vec — Function

rand_prob_vec(N)
rand_prob_vec(rng, N)

Generate a random probability distribution of size N.

HiddenMarkovModels.rand_trans_mat — Function

rand_trans_mat(N)
rand_trans_mat(rng, N)

Generate a random transition matrix of size (N, N).

HiddenMarkovModels.fit_element_from_sequence! — Function

fit_element_from_sequence!(dists, i, x, w)

Modify the i-th element of dists by fitting it to an observation sequence x with associated weight sequence w.

The default behavior is a fallback on StatsAPI.fit!, which users are encouraged to implement if their observation distributions are mutable. If this is not possible, please override fit_element_from_sequence! directly.

HiddenMarkovModels.LightDiagNormal — Type

struct LightDiagNormal{T1, T2, T3, V1<:AbstractArray{T1, 1}, V2<:AbstractArray{T2, 1}, V3<:AbstractArray{T3, 1}}

An HMMs-compatible implementation of a multivariate normal distribution with diagonal covariance, enabling allocation-free estimation.

This is not part of the public API and is expected to change.

Fields

μ::AbstractVector: vector of means

σ::AbstractVector: vector of standard deviations

logσ::AbstractVector: vector of log standard deviations

In-place algorithms (internals)

Storage types

HiddenMarkovModels.ForwardStorage — Type

struct ForwardStorage{R}

Store forward quantities with element type R.

This storage is relative to a single sequence.

Fields

The only fields useful outside of the algorithm are α and logL, the rest does not belong to the public API.

logL::Base.RefValue: total loglikelihood

logb::Vector: observation loglikelihoods logbₜ[i] = ℙ(Y[t] | X[t]=i)

α::Vector: scaled forward messsages for a given time step

α_next::Vector: same as α but for the next time step

HiddenMarkovModels.ViterbiStorage — Type

struct ViterbiStorage{R}

Store Viterbi quantities with element type R.

This storage is relative to a single sequence.

Fields

The only field useful outside of the algorithm is q, the rest does not belong to the public API.

logb::Vector: observation loglikelihoods at a given time step

δ::Vector: highest path scores when accounting for the first t observations and ending at a given state

δ_prev::Vector: same as δ but for the previous time step

ψ::Matrix{Int64}: penultimate state maximizing the path score

q::Vector{Int64}: most likely state at each time q[t] = argmaxᵢ ℙ(X[t]=i | Y[1:T])

scratch::Vector: scratch storage space

HiddenMarkovModels.ForwardBackwardStorage — Type

struct ForwardBackwardStorage{R, M<:AbstractArray{R, 2}}

Store forward-backward quantities with element type R.

This storage is relative to a single sequence.

Fields

The only fields useful outside of the algorithm are γ, logL, init_count and trans_count, the rest does not belong to the public API.

logL::Base.RefValue: total loglikelihood

α::Matrix: scaled forward messsages α[i,t] proportional to ℙ(Y[1:t], X[t]=i) (up to a function of t)

β::Matrix: scaled backward messsages β[i,t] proportional to ℙ(Y[t+1:T] | X[t]=i) (up to a function of t)

γ::Matrix: posterior state marginals γ[i,t] = ℙ(X[t]=i | Y[1:T])

ξ::Vector{M} where {R, M<:AbstractMatrix{R}}: posterior transition marginals ξ[t][i,j] = ℙ(X[t]=i, X[t+1]=j | Y[1:T])

c::Vector: forward message inverse normalizations c[t] = 1 / sum(α[:,t])

logB::Matrix: observation loglikelihoods logB[i,t] = ℙ(Y[t] | X[t]=i)

logm::Vector: maximum of the observation loglikelihoods logm[t] = maximum(logB[:, t])

B::Matrix: numerically stabilized observation likelihoods B[i,t] = exp.(logB[i,t] - logm[t])

Bβ::Matrix: product Bβ[i,t] = B[i,t] * β[i,t]

HiddenMarkovModels.BaumWelchStorage — Type

struct BaumWelchStorage{R, M<:AbstractArray{R, 2}}

Store Baum-Welch quantities with element type R.

This storage is relative to a several sequences.

Fields

These fields (except limits) are passed to the fit!(hmm, ...) method along with obs_seqs_concat.

init_count::Vector: posterior initialization counts for each state

trans_count::AbstractMatrix: posterior transition counts for each state

state_marginals_concat::Matrix: concatenation along time of the state marginal matrices γ[i,t] = ℙ(X[t]=i | Y[1:T]) for all observation sequences

limits::Vector{Int64}: temporal separations between observation sequences: state_marginals_concat[limits[k]+1:limits[k+1]] refers to sequence k

Initializing storage

HiddenMarkovModels.initialize_forward — Function

initialize_forward(hmm, obs_seq)

HiddenMarkovModels.initialize_viterbi — Function

initialize_viterbi(hmm, obs_seq)

HiddenMarkovModels.initialize_forward_backward — Function

initialize_forward_backward(hmm, obs_seq)

HiddenMarkovModels.initialize_baum_welch — Function

initialize_baum_welch(hmm, obs_seqs, nb_seqs)

HiddenMarkovModels.initialize_logL_evolution — Function

initialize_logL_evolution(hmm, obs_seqs, nb_seqs; max_iterations)

Modifying storage

HiddenMarkovModels.forward! — Function

forward!(storage, hmm, obs_seq)

HiddenMarkovModels.viterbi! — Function

viterbi!(storage, hmm, obs_seq)

HiddenMarkovModels.forward_backward! — Function

forward_backward!(storage, hmm, obs_seq)

HiddenMarkovModels.baum_welch! — Function

baum_welch!(
    fb_storages, bw_storage, logL_evolution,
    hmm, obs_seqs, obs_seqs_concat;
    atol, max_iterations, loglikelihood_increasing
)

Notations

Integers

N: number of states

D: dimension of the observations

T: trajectory length

K: number of trajectories

Models and simulations

p or init: initialization (vector of state probabilities)

A or trans: transition_matrix (matrix of transition probabilities)

d or dists: observation distribution (vector of rand-able and logdensityof-able objects)

state_seq: a sequence of states (vector of integers)

obs_seq: a sequence of observations (vector of individual observations)

obs_seqs: several sequences of observations

nb_seqs: number of observation sequences

Forward backward

(log)b: vector of observation (log)likelihoods by state for an individual observation

(log)B: matrix of observation (log)likelihoods by state for a sequence of observations

α: scaled forward variables

β: scaled backward variables

γ: state marginals

ξ: transition marginals

logL: posterior loglikelihood of a sequence of observations

Index

HiddenMarkovModels

HiddenMarkovModels.AbstractHMM

HiddenMarkovModels.AbstractHiddenMarkovModel

HiddenMarkovModels.BaumWelchStorage

HiddenMarkovModels.ForwardBackwardStorage

HiddenMarkovModels.ForwardStorage

HiddenMarkovModels.HMM

HiddenMarkovModels.HiddenMarkovModel

HiddenMarkovModels.LightDiagNormal

HiddenMarkovModels.ViterbiStorage

Base.eltype

Base.length

Base.rand

DensityInterface.logdensityof

HiddenMarkovModels.baum_welch

HiddenMarkovModels.baum_welch!

HiddenMarkovModels.fit_element_from_sequence!

HiddenMarkovModels.forward

HiddenMarkovModels.forward!

HiddenMarkovModels.forward_backward

HiddenMarkovModels.forward_backward!

HiddenMarkovModels.initialization

HiddenMarkovModels.initialize_baum_welch

HiddenMarkovModels.initialize_forward

HiddenMarkovModels.initialize_forward_backward

HiddenMarkovModels.initialize_logL_evolution

HiddenMarkovModels.initialize_viterbi

HiddenMarkovModels.obs_distributions

HiddenMarkovModels.rand_prob_vec

HiddenMarkovModels.rand_trans_mat

HiddenMarkovModels.transition_matrix

HiddenMarkovModels.viterbi

HiddenMarkovModels.viterbi!

StatsAPI.fit!