API reference

HiddenMarkovModels — Module

HiddenMarkovModels

A Julia package for HMM modeling, simulation, inference and learning.

HiddenMarkovModels.HMMs — Module

HMMs

Alias for the module HiddenMarkovModels.

Types

HiddenMarkovModels.AbstractHiddenMarkovModel — Type

AbstractHiddenMarkovModel

Interface for an HMM amenable to simulation, inference and learning.

Required interface

initial_distribution(hmm)

transition_matrix(hmm)

obs_distribution(hmm, i)

fit!(hmm, init_count, trans_count, obs_seq, state_marginals) (optional)

Applicable methods

rand([rng,] hmm, T)

logdensityof(hmm, obs_seq)

viterbi(hmm, obs_seq)

forward_backward(hmm, obs_seq)

baum_welch(hmm, obs_seq) (if fit! is implemented)

HiddenMarkovModels.HiddenMarkovModel — Type

HiddenMarkovModel{D} <: AbstractHMM

Basic implementation of an HMM.

Fields

init::AbstractVector: initial state probabilities

trans::AbstractMatrix: state transition matrix

dists::AbstractVector{D}: observation distributions

Constructors

HMM(init, trans, dists)

HiddenMarkovModels.AbstractHMM — Type

AbstractHMM

Alias for the type AbstractHiddenMarkovModel.

HiddenMarkovModels.HMM — Type

HMM

Alias for the type HiddenMarkovModel.

Basics

Base.length — Function

length(hmm::AbstractHMM)

Return the number of states of hmm.

Base.rand — Function

rand(rng, hmm, T)

Simulate hmm for T time steps with a specified rng.

rand(hmm::AbstractHMM, T)

Simulate hmm for T time steps with the global RNG.

HiddenMarkovModels.initial_distribution — Function

initial_distribution(hmm::AbstractHMM)

Return the initial state probabilities of hmm.

HiddenMarkovModels.transition_matrix — Function

transition_matrix(hmm::AbstractHMM)

Return the state transition probabilities of hmm.

HiddenMarkovModels.obs_distribution — Function

obs_distribution(hmm::AbstractHMM, i)

Return the observation distribution of hmm associated with state i.

The returned object dist must implement

rand(rng, dist)

DensityInterface.logdensityof(dist, x)

Inference

DensityInterface.logdensityof — Function

DensityInterface.logdensityof(hmm, obs_seq)

Apply the forward algorithm to compute the loglikelihood of a single observation sequence for an HMM.

DensityInterface.logdensityof(hmm, obs_seqs, nb_seqs)

Apply the forward algorithm to compute the total loglikelihood of multiple observation sequences for an HMM

HiddenMarkovModels.viterbi — Function

viterbi(hmm, obs_seq)

Apply the Viterbi algorithm to compute the most likely state sequence of an HMM for a single observation sequence.

viterbi(hmm, obs_seqs, nb_seqs)

Apply the Viterbi algorithm to compute the most likely state sequences of an HMM for multiple observation sequences.

HiddenMarkovModels.forward_backward — Function

forward_backward(hmm, obs_seq)

Apply the forward-backward algorithm to estimate the posterior state marginals of an HMM for a single observation sequence, and return a ForwardBackwardStorage.

forward_backward(hmm, obs_seqs, nb_seqs)

Apply the forward-backward algorithm to estimate the posterior state marginals of an HMM for multiple observation sequences, and return a ForwardBackwardStorage.

HiddenMarkovModels.ForwardBackwardStorage — Type

ForwardBackwardStorage{R}

Store forward-backward quantities with element type R.

Fields

α::Matrix{R}: forward variables

β::Matrix{R}: backward variables

γ::Matrix{R}: one-state marginals

ξ::Array{R,3}: two-state marginals

_c::Vector{R}: forward variable inverse normalizations

_m::Vector{R}: maximum of logB

_Bβ::Matrix{R}: stabilized product Bβ

Learning

HiddenMarkovModels.baum_welch — Function

baum_welch(
    hmm_init, obs_seq;
    max_iterations, rtol, check_loglikelihood_increasing
)

Apply the Baum-Welch algorithm to estimate the parameters of an HMM and return a tuple (hmm, logL_evolution).

The procedure is based on a single observation sequence and initialized with hmm_init.

baum_welch(
    hmm_init, obs_seqs, nb_seqs;
    max_iterations, rtol, check_loglikelihood_increasing
)

Apply the Baum-Welch algorithm to estimate the parameters of an HMM and return a tuple (hmm, logL_evolution).

The procedure is based on multiple observation sequences and initialized with hmm_init.

Reimplement if needed

StatsAPI.fit! — Function

StatsAPI.fit!(hmm::AbstractHMM, init_count, trans_count, obs_seq, state_marginals)

Update hmm in-place based on information generated during forward-backward.

HiddenMarkovModels.fit_element_from_sequence! — Function

fit_element_from_sequence!(dists, i, x, w)

Modify the i-th element of dists by fitting it to an observation sequence x with associated weight sequence w.

The default behavior is a fallback on StatsAPI.fit!(dists[i], x, w), which users are encouraged to implement if their observation distributions are mutable. If not, check out the source code below to see the hack we used for Distributions.jl, and imitate it.

HiddenMarkovModels.LightDiagNormal — Type

LightDiagNormal

An HMMs-compatible implementation of a multivariate normal distribution with diagonal covariance, enabling allocation-free estimation.

This is not part of the public API and is expected to change.

Index

HiddenMarkovModels

HiddenMarkovModels.HMMs

HiddenMarkovModels.AbstractHMM

HiddenMarkovModels.AbstractHiddenMarkovModel

HiddenMarkovModels.ForwardBackwardStorage

HiddenMarkovModels.HMM

HiddenMarkovModels.HiddenMarkovModel

HiddenMarkovModels.LightDiagNormal

Base.length

Base.rand

DensityInterface.logdensityof

HiddenMarkovModels.baum_welch

HiddenMarkovModels.fit_element_from_sequence!

HiddenMarkovModels.forward_backward

HiddenMarkovModels.initial_distribution

HiddenMarkovModels.obs_distribution

HiddenMarkovModels.transition_matrix

HiddenMarkovModels.viterbi

StatsAPI.fit!

Notations

Integers

N: number of states

D: dimension of the observations

T: trajectory length

K: number of trajectories

Models and simulations

p or init: initial_distribution (vector of state probabilities)

A or trans: transition_matrix (matrix of transition probabilities)

dists: observation distribution (vector of rand-able and logdensityof-able objects)

state_seq: a sequence of states (vector of integers)

obs_seq: a sequence of observations (vector of individual observations)

obs_seqs: several sequences of observations

Forward backward

(log)b: vector of observation (log)likelihoods by state for an individual observation

(log)B: matrix of observation (log)likelihoods by state for a sequence of observations

α: forward variables

β: backward variables

γ: one-state marginals

ξ: two-state marginals

logL: loglikelihood of a sequence of observations