API reference

HiddenMarkovModels — Module

HiddenMarkovModels

A Julia package for HMM modeling, simulation, inference and learning.

HiddenMarkovModels.HMMs — Module

HMMs

Alias for the module HiddenMarkovModels.

Types

HiddenMarkovModels.AbstractModel — Type

AbstractModel

Abstract supertype for AbstractMarkovChain and AbstractHiddenMarkovModel.

Not exported.

Markov chains

HiddenMarkovModels.AbstractMarkovChain — Type

AbstractMarkovChain <: AbstractModel

Abstract supertype for a Markov chain amenable to simulation, inference and learning.

Required interface

initial_distribution(mc)

transition_matrix(mc)

fit!(mc, init_count, trans_count) (optional)

Applicable methods

rand([rng,] mc, T)

logdensityof(mc, state_seq)

fit(mc, state_seq_or_seqs) (if fit! is implemented)

HiddenMarkovModels.MarkovChain — Type

MarkovChain <: AbstractMarkovChain

Basic implementation of a discrete-state Markov chain.

Fields

init::AbstractVector: initial state probabilities

trans::AbstractMatrix: state transition matrix

HiddenMarkovModels.AbstractMC — Type

AbstractMC

Alias for the type AbstractMarkovChain.

HiddenMarkovModels.MC — Type

MC

Alias for the type MarkovChain.

Hidden Markov Models

HiddenMarkovModels.AbstractHiddenMarkovModel — Type

AbstractHiddenMarkovModel <: AbstractModel

Abstract supertype for an HMM amenable to simulation, inference and learning.

Required interface

initial_distribution(hmm)

transition_matrix(hmm)

obs_distribution(hmm, i)

fit!(hmm, init_count, trans_count, obs_seq, state_marginals) (optional)

Applicable methods

rand([rng,] hmm, T)

logdensityof(hmm, obs_seq) / logdensityof(hmm, obs_seqs, nb_seqs)

viterbi(hmm, obs_seq) / viterbi(hmm, obs_seqs, nb_seqs)

forward_backward(hmm, obs_seq) / forward_backward(hmm, obs_seqs, nb_seqs)

baum_welch(hmm, obs_seq) / baum_welch(hmm, obs_seqs, nb_seqs) if fit! is implemented

HiddenMarkovModels.HiddenMarkovModel — Type

HiddenMarkovModel{D} <: AbstractHiddenMarkovModel

Basic implementation of an HMM.

Fields

init::AbstractVector: initial state probabilities

trans::AbstractMatrix: state transition matrix

dists::AbstractVector{D}: observation distributions

HiddenMarkovModels.AbstractHMM — Type

AbstractHMM

Alias for the type AbstractHiddenMarkovModel.

HiddenMarkovModels.HMM — Type

HMM

Alias for the type HiddenMarkovModel.

Basics

Base.rand — Function

rand([rng=default_rng(),] model::AbstractModel, T)

Simulate model for T time steps with a specified rng.

Base.length — Function

length(model::AbstractModel)

Return the number of states of model.

HiddenMarkovModels.initial_distribution — Function

initial_distribution(model::AbstractModel)

Return the initial state probabilities of model.

HiddenMarkovModels.transition_matrix — Function

transition_matrix(model::AbstractModel)

Return the state transition probabilities of model.

HiddenMarkovModels.obs_distribution — Function

obs_distribution(hmm::AbstractHMM, i)

Return the observation distribution of hmm associated with state i.

The returned object dist must implement

rand(rng, dist)

DensityInterface.logdensityof(dist, x)

Inference

DensityInterface.logdensityof — Function

logdensityof(mc, state_seq)

Compute the loglikelihood of a single state sequence for a Markov chain.

logdensityof(hmm, obs_seq)

Apply the forward algorithm to compute the loglikelihood of a single observation sequence for an HMM.

logdensityof(hmm, obs_seqs, nb_seqs)

Apply the forward algorithm to compute the total loglikelihood of multiple observation sequences for an HMM.

Multithreading

This function is parallelized across sequences.

HiddenMarkovModels.viterbi — Function

viterbi(hmm, obs_seq)

Apply the Viterbi algorithm to compute the most likely state sequence of an HMM for a single observation sequence.

viterbi(hmm, obs_seqs, nb_seqs)

Apply the Viterbi algorithm to compute the most likely state sequences of an HMM for multiple observation sequences.

Multithreading

This function is parallelized across sequences.

HiddenMarkovModels.forward_backward — Function

forward_backward(hmm, obs_seq)

Apply the forward-backward algorithm to estimate the posterior state marginals of an HMM for a single observation sequence, and return a ForwardBackwardStorage.

forward_backward(hmm, obs_seqs, nb_seqs)

Apply the forward-backward algorithm to estimate the posterior state marginals of an HMM for multiple observation sequences, and return a ForwardBackwardStorage.

Multithreading

This function is parallelized across sequences.

Learning

StatsAPI.fit! — Function

fit!(mc::MC, init_count, trans_count)

Update mc in-place based on information generated from a state sequence.

fit!(hmm::HMM, init_count, trans_count, obs_seq, state_marginals)

Update hmm in-place based on information generated during forward-backward.

StatsAPI.fit — Function

fit(mc, state_seq_or_seqs)

Fit a Markov chain of the same type as mc to one or several state sequence(s).

Beware that mc must be an actual object of type MarkovChain, and not the type itself as is usually done eg. in Distributions.jl.

HiddenMarkovModels.baum_welch — Function

baum_welch(
    hmm_init, obs_seq;
    max_iterations, rtol, check_loglikelihood_increasing
)

Apply the Baum-Welch algorithm to estimate the parameters of an HMM and return a tuple (hmm, logL_evolution).

The procedure is based on a single observation sequence and initialized with hmm_init.

Error

Using a relative tolerance rtol to track convergence makes no sense because the loglikelihood can change signs. Always set rtol=-Inf for now, this will be fixed in the next breaking release.

baum_welch(
    hmm_init, obs_seqs, nb_seqs;
    max_iterations, rtol, check_loglikelihood_increasing
)

Apply the Baum-Welch algorithm to estimate the parameters of an HMM and return a tuple (hmm, logL_evolution).

The procedure is based on multiple observation sequences and initialized with hmm_init.

Multithreading

This function is parallelized across sequences.

Error

Using a relative tolerance rtol to track convergence makes no sense because the loglikelihood can change signs. Always set rtol=-Inf for now, this will be fixed in the next breaking release.

Internals

HiddenMarkovModels.ForwardBackwardStorage — Type

ForwardBackwardStorage{R}

Store forward-backward quantities with element type R.

Fields

α::Matrix{R}: forward variables

β::Matrix{R}: backward variables

γ::Matrix{R}: one-state marginals

ξ::Array{R,3}: two-state marginals

_c::Vector{R}: forward variable inverse normalizations

_m::Vector{R}: maximum of logB

_Bβ::Matrix{R}: stabilized product Bβ

HiddenMarkovModels.fit_element_from_sequence! — Function

fit_element_from_sequence!(dists, i, x, w)

Modify the i-th element of dists by fitting it to an observation sequence x with associated weight sequence w.

The default behavior is a fallback on StatsAPI.fit!(dists[i], x, w), which users are encouraged to implement if their observation distributions are mutable. If not, check out the source code below to see the hack we used for Distributions.jl, and imitate it.

HiddenMarkovModels.LightDiagNormal — Type

LightDiagNormal

An HMMs-compatible implementation of a multivariate normal distribution with diagonal covariance, enabling allocation-free estimation.

This is not part of the public API and is expected to change.

Notations

Integers

N: number of states

D: dimension of the observations

T: trajectory length

K: number of trajectories

Models and simulations

p or init: initial_distribution (vector of state probabilities)

A or trans: transition_matrix (matrix of transition probabilities)

dists: observation distribution (vector of rand-able and logdensityof-able objects)

state_seq: a sequence of states (vector of integers)

obs_seq: a sequence of observations (vector of individual observations)

obs_seqs: several sequences of observations

Forward backward

(log)b: vector of observation (log)likelihoods by state for an individual observation

(log)B: matrix of observation (log)likelihoods by state for a sequence of observations

α: forward variables

β: backward variables

γ: one-state marginals

ξ: two-state marginals

logL: loglikelihood of a sequence of observations

Index

HiddenMarkovModels

HiddenMarkovModels.HMMs

HiddenMarkovModels.AbstractHMM

HiddenMarkovModels.AbstractHiddenMarkovModel

HiddenMarkovModels.AbstractMC

HiddenMarkovModels.AbstractMarkovChain

HiddenMarkovModels.AbstractModel

HiddenMarkovModels.ForwardBackwardStorage

HiddenMarkovModels.HMM

HiddenMarkovModels.HiddenMarkovModel

HiddenMarkovModels.LightDiagNormal

HiddenMarkovModels.MC

HiddenMarkovModels.MarkovChain

Base.length

Base.rand

DensityInterface.logdensityof

HiddenMarkovModels.baum_welch

HiddenMarkovModels.fit_element_from_sequence!

HiddenMarkovModels.forward_backward

HiddenMarkovModels.initial_distribution

HiddenMarkovModels.obs_distribution

HiddenMarkovModels.transition_matrix

HiddenMarkovModels.viterbi

StatsAPI.fit

StatsAPI.fit!