API reference

HiddenMarkovModels — Module

HiddenMarkovModels

A Julia package for HMM modeling, simulation, inference and learning.

HiddenMarkovModels.HMMs — Module

HMMs

Alias for the module HiddenMarkovModels.

Types

Markov chains

HiddenMarkovModels.AbstractMarkovChain — Type

AbstractMarkovChain

Abstract supertype for a Markov chain amenable to simulation, inference and learning.

Required interface

initial_distribution(mc)

transition_matrix(mc)

fit!(mc, init_count, trans_count) (optional)

Applicable methods

rand([rng,] mc, T)

logdensityof(mc, state_seq)

fit(mc, state_seq_or_seqs) (if fit! is implemented)

HiddenMarkovModels.MarkovChain — Type

MarkovChain <: AbstractMarkovChain

Basic implementation of a discrete-state Markov chain.

Fields

init::AbstractVector: initial state probabilities

trans::AbstractMatrix: state transition matrix

HiddenMarkovModels.AbstractMC — Type

AbstractMC

Alias for the type AbstractMarkovChain.

HiddenMarkovModels.MC — Type

MC

Alias for the type MarkovChain.

Hidden Markov Models

HiddenMarkovModels.AbstractHiddenMarkovModel — Type

AbstractHiddenMarkovModel <: AbstractMarkovChain

Abstract supertype for an HMM amenable to simulation, inference and learning.

Required interface

initial_distribution(hmm)

transition_matrix(hmm)

obs_distribution(hmm, i)

fit!(hmm, init_count, trans_count, obs_seq, state_marginals) (optional)

Applicable methods

rand([rng,] hmm, T)

logdensityof(hmm, obs_seq) / logdensityof(hmm, obs_seqs, nb_seqs)

forward(hmm, obs_seq) / forward(hmm, obs_seqs, nb_seqs)

viterbi(hmm, obs_seq) / viterbi(hmm, obs_seqs, nb_seqs)

forward_backward(hmm, obs_seq) / forward_backward(hmm, obs_seqs, nb_seqs)

baum_welch(hmm, obs_seq) / baum_welch(hmm, obs_seqs, nb_seqs) if fit! is implemented

HiddenMarkovModels.HiddenMarkovModel — Type

HiddenMarkovModel{D} <: AbstractHiddenMarkovModel

Basic implementation of an HMM.

Fields

init::AbstractVector: initial state probabilities

trans::AbstractMatrix: state transition matrix

dists::AbstractVector{D}: observation distributions

HiddenMarkovModels.AbstractHMM — Type

AbstractHMM

Alias for the type AbstractHiddenMarkovModel.

HiddenMarkovModels.HMM — Type

HMM

Alias for the type HiddenMarkovModel.

Basics

Base.rand — Function

rand([rng=default_rng(),] mc::AbstractMarkovChain, T)

Simulate mc for T time steps with a specified rng.

Base.length — Function

length(mc::AbstractMarkovChain)

Return the number of states of model.

HiddenMarkovModels.initial_distribution — Function

initial_distribution(mc::AbstractMarkovChain)

Return the initial state probabilities of mc.

HiddenMarkovModels.transition_matrix — Function

transition_matrix(mc::AbstractMarkovChain)

Return the state transition probabilities of mc.

HiddenMarkovModels.obs_distribution — Function

obs_distribution(hmm::AbstractHMM, i)

Return the observation distribution of hmm associated with state i.

The returned object dist must implement

rand(rng, dist)

DensityInterface.logdensityof(dist, x)

Inference

DensityInterface.logdensityof — Function

logdensityof(mc, state_seq)

Compute the loglikelihood of a single state sequence for a Markov chain.

logdensityof(hmm, obs_seq)

Apply the forward algorithm to compute the loglikelihood of a single observation sequence for an HMM.

Return a number.

logdensityof(hmm, obs_seqs, nb_seqs)

Apply the forward algorithm to compute the total loglikelihood of multiple observation sequences for an HMM.

Return a number.

Multithreading

This function is parallelized across sequences.

HiddenMarkovModels.forward — Function

forward(hmm, obs_seq)

Apply the forward algorithm to an HMM.

Return a tuple (α, logL) where

logL is the loglikelihood of the sequence

α[i] is the posterior probability of state i at the end of the sequence.

forward(hmm, obs_seqs, nb_seqs)

Apply the forward algorithm to an HMM, based on multiple observation sequences.

Return a vector of tuples (αₖ, logLₖ), where

logLₖ is the loglikelihood of sequence k

αₖ[i] is the posterior probability of state i at the end of sequence k

Multithreading

This function is parallelized across sequences.

HiddenMarkovModels.viterbi — Function

viterbi(hmm, obs_seq)

Apply the Viterbi algorithm to compute the most likely state sequence of an HMM.

Return a vector of integers.

viterbi(hmm, obs_seqs, nb_seqs)

Apply the Viterbi algorithm to compute the most likely state sequences of an HMM, based on multiple observation sequences.

Return a vector of vectors of integers.

Multithreading

This function is parallelized across sequences.

HiddenMarkovModels.forward_backward — Function

forward_backward(hmm, obs_seq)

Apply the forward-backward algorithm to estimate the posterior state marginals of an HMM.

Return a ForwardBackwardStorage.

forward_backward(hmm, obs_seqs, nb_seqs)

Apply the forward-backward algorithm to estimate the posterior state marginals of an HMM, based on multiple observation sequences.

Return a vector of ForwardBackwardStorage objects.

Multithreading

This function is parallelized across sequences.

Learning

StatsAPI.fit! — Function

fit!(mc::MC, init_count, trans_count)

Update mc in-place based on information generated from a state sequence.

fit!(hmm::HMM, init_count, trans_count, obs_seq, state_marginals)

Update hmm in-place based on information generated during forward-backward.

StatsAPI.fit — Function

fit(mc, state_seq_or_seqs)

Fit a Markov chain of the same type as mc to one or several state sequence(s).

Beware that mc must be an actual object of type MarkovChain, and not the type itself as is usually done eg. in Distributions.jl.

HiddenMarkovModels.baum_welch — Function

baum_welch(
    hmm_init, obs_seq;
    atol, max_iterations, check_loglikelihood_increasing
)

Apply the Baum-Welch algorithm to estimate the parameters of an HMM starting from hmm_init.

Return a tuple (hmm_est, logL_evolution).

Keyword arguments

atol: Minimum loglikelihood increase at an iteration of the algorithm (otherwise the algorithm is deemed to have converged)

max_iterations: Maximum number of iterations of the algorithm

check_loglikelihood_increasing: Whether to throw an error if the loglikelihood decreases

baum_welch(
    hmm_init, obs_seqs, nb_seqs;
    atol, max_iterations, check_loglikelihood_increasing
)

Apply the Baum-Welch algorithm to estimate the parameters of an HMM starting from hmm_init, based on nb_seqs observation sequences.

Return a tuple (hmm_est, logL_evolution).

Multithreading

This function is parallelized across sequences.

Keyword arguments

atol: Minimum loglikelihood increase at an iteration of the algorithm (otherwise the algorithm is deemed to have converged)

max_iterations: Maximum number of iterations of the algorithm

check_loglikelihood_increasing: Whether to throw an error if the loglikelihood decreases

Internals

HiddenMarkovModels.ForwardBackwardStorage — Type

ForwardBackwardStorage{R}

Store forward-backward quantities with element type R.

Fields

Let X denote the vector of hidden states and Y denote the vector of observations. The following fields are part of the API:

γ::Matrix{R}: posterior one-state marginals γ[i,t] = ℙ(X[t]=i | Y[1:T])

ξ::Array{R,3}: posterior two-state marginals ξ[i,j,t] = ℙ(X[t:t+1]=(i,j) | Y[1:T])

The following fields are internals and subject to change:

α::Matrix{R}: scaled forward variables α[i,t] proportional to ℙ(Y[1:t], X[t]=i) (up to a function of t)

β::Matrix{R}: scaled backward variables β[i,t] proportional to ℙ(Y[t+1:T] | X[t]=i) (up to a function of t)

c::Vector{R}: forward variable inverse normalizations c[t] = 1 / sum(α[:, t])

logm::Vector{R}: maximum of the observation loglikelihoods logB

Bscaled::Matrix{R}: numerically stabilized observation likelihoods B

Bβscaled::Matrix{R}: numerically stabilized product Bβ

HiddenMarkovModels.fit_element_from_sequence! — Function

fit_element_from_sequence!(dists, i, x, w)

Modify the i-th element of dists by fitting it to an observation sequence x with associated weight sequence w.

The default behavior is a fallback on StatsAPI.fit!, which users are encouraged to implement if their observation distributions are mutable. If this is not possible, please override fit_element_from_sequence! directly.

HiddenMarkovModels.LightDiagNormal — Type

LightDiagNormal

An HMMs-compatible implementation of a multivariate normal distribution with diagonal covariance, enabling allocation-free estimation.

This is not part of the public API and is expected to change.

Notations

Integers

N: number of states

D: dimension of the observations

T: trajectory length

K: number of trajectories

Models and simulations

p or init: initial_distribution (vector of state probabilities)

A or trans: transition_matrix (matrix of transition probabilities)

dists: observation distribution (vector of rand-able and logdensityof-able objects)

state_seq: a sequence of states (vector of integers)

obs_seq: a sequence of observations (vector of individual observations)

obs_seqs: several sequences of observations

Forward backward

(log)b: vector of observation (log)likelihoods by state for an individual observation

(log)B: matrix of observation (log)likelihoods by state for a sequence of observations

α: scaled forward variables

β: scaled backward variables

γ: one-state marginals

ξ: two-state marginals

logL: loglikelihood of a sequence of observations

Index

HiddenMarkovModels

HiddenMarkovModels.HMMs

HiddenMarkovModels.AbstractHMM

HiddenMarkovModels.AbstractHiddenMarkovModel

HiddenMarkovModels.AbstractMC

HiddenMarkovModels.AbstractMarkovChain

HiddenMarkovModels.ForwardBackwardStorage

HiddenMarkovModels.HMM

HiddenMarkovModels.HiddenMarkovModel

HiddenMarkovModels.LightDiagNormal

HiddenMarkovModels.MC

HiddenMarkovModels.MarkovChain

Base.length

Base.rand

DensityInterface.logdensityof

HiddenMarkovModels.baum_welch

HiddenMarkovModels.fit_element_from_sequence!

HiddenMarkovModels.forward

HiddenMarkovModels.forward_backward

HiddenMarkovModels.initial_distribution

HiddenMarkovModels.obs_distribution

HiddenMarkovModels.transition_matrix

HiddenMarkovModels.viterbi

StatsAPI.fit

StatsAPI.fit!