API reference

HiddenMarkovModels — Module

HiddenMarkovModels

A Julia package for HMM modeling, simulation, inference and learning.

HiddenMarkovModels.HMMs — Module

HMMs

Alias for the module HiddenMarkovModels.

Types

HiddenMarkovModels.HiddenMarkovModel — Type

HiddenMarkovModel{SP<:StateProcess,OP<:ObservationProcess}

Combination of a state and an observation process, amenable to simulation, inference and learning.

Fields

state_process::SP

obs_process::OP

Both fields are considered part of the API, which is why there are no accessors.

Applicable methods

initial_distribution(hmm)

transition_matrix(hmm)

obs_distribution(hmm, i)

rand([rng,] hmm, T)

logdensityof(hmm, obs_seq)

viterbi(hmm, obs_seq)

forward_backward(hmm, obs_seq)

baum_welch(hmm, obs_seq)

HiddenMarkovModels.HMM — Type

HMM

Alias for the struct HiddenMarkovModel.

HiddenMarkovModels.StateProcess — Type

StateProcess

Abstract type for the state part of an HMM.

Required methods

length(sp)

initial_distribution(sp)

transition_matrix(sp)

Optional methods

fit!(sp, p_count, A_count)

HiddenMarkovModels.StandardStateProcess — Type

StandardStateProcess <: StateProcess

Fields

p::AbstractVector: initial distribution

A::AbstractMatrix: transition matrix

HiddenMarkovModels.ObservationProcess — Type

ObservationProcess

Abstract type for the observation part of an HMM.

Required methods

length(op)

obs_distribution(op, i)

Optional methods

fit!(op, obs_seq, γ)

HiddenMarkovModels.StandardObservationProcess — Type

StandardObservationProcess{D} <: ObservationProcess

Fields

distributions::AbstractVector{D}: one distribution per state

Basics

Base.length — Function

length(sp::StateProcess)

Return the number of states of sp.

length(op::ObservationProcess)

Return the number of states of op.

length(hmm::HMM)

Return the number of states of hmm.

Base.rand — Function

rand(rng, hmm, T)

Simulate an HMM for T time steps with a specified rng.

rand(hmm, T)

Simulate an HMM for T time steps.

HiddenMarkovModels.initial_distribution — Function

initial_distribution(sp::StateProcess)

Return the initial state probabilities of sp.

initial_distribution(hmm::HMM)

Return the initial state probabilities of hmm.

HiddenMarkovModels.transition_matrix — Function

transition_matrix(sp::StateProcess)

Return the state transition probabilities of sp.

transition_matrix(hmm::HMM)

Return the state transition probabilities of hmm.

HiddenMarkovModels.obs_distribution — Function

obs_distribution(op::ObservationProcess, i)

Return the observation distribution of op associated with state i.

obs_distribution(hmm::HMM, i)

Return the observation distribution of hmm associated with state i.

Inference

DensityInterface.logdensityof — Function

DensityInterface.logdensityof(hmm, obs_seq)

Apply the forward algorithm to compute the loglikelihood of a single observation sequence for an HMM.

DensityInterface.logdensityof(hmm, obs_seqs, nb_seqs)

Apply the forward algorithm to compute the total loglikelihood of multiple observation sequences for an HMM

HiddenMarkovModels.viterbi — Function

viterbi(hmm, obs_seq)

Apply the Viterbi algorithm to compute the most likely state sequence of an HMM for a single observation sequence.

viterbi(hmm, obs_seqs, nb_seqs)

Apply the Viterbi algorithm to compute the most likely state sequences of an HMM for multiple observation sequences.

HiddenMarkovModels.forward_backward — Function

forward_backward(hmm, obs_seq)

Apply the forward-backward algorithm to estimate the posterior state marginals of an HMM for a single observation sequence.

forward_backward(hmm, obs_seqs, nb_seqs)

Apply the forward-backward algorithm to estimate the posterior state marginals of an HMM for multiple observation sequences.

Learning

HiddenMarkovModels.baum_welch — Function

baum_welch(hmm_init, obs_seq; max_iterations, rtol)

Apply the Baum-Welch algorithm to estimate the parameters of an HMM based on a single observation sequence, and return a tuple (hmm, logL_evolution).

baum_welch(hmm_init, obs_seqs, nb_seqs; max_iterations, rtol)

Apply the Baum-Welch algorithm to estimate the parameters of an HMM based on multiple observation sequences, and return a tuple (hmm, logL_evolution).

Reimplement if needed

StatsAPI.fit! — Function

StatsAPI.fit!(op::ObservationProcess, obs_seq, γ)

Update the initial distribution and transition matrix of sp based on weighted initialization counts p_count and transition counts A_count.

StatsAPI.fit!(op::ObservationProcess, obs_seq, γ)

Update all observation distributions of op based on an observation sequence obs_seq, weighted by γ[i, :] for each state i.

HiddenMarkovModels.fit_element_from_sequence! — Function

fit_element_from_sequence!(dists, i, x, w)

Modify the i-th element of dists by fitting it to an observation sequence x with associated weight sequence w.

The default behavior is a fallback on StatsAPI.fit!(dists[i], x, w), which users are encouraged to implement if their observation distributions are mutable. If not, check out the source code to see the hack we used for Distributions.jl, and imitate it.

HiddenMarkovModels.LightDiagNormal — Type

LightDiagonalNormal

An HMMs-compatible implementation of a multivariate normal distribution with diagonal covariance, enabling allocation-free estimation.

This is not part of the public API and is expected to change.

Index

HiddenMarkovModels

HiddenMarkovModels.HMMs

HiddenMarkovModels.HMM

HiddenMarkovModels.HiddenMarkovModel

HiddenMarkovModels.LightDiagNormal

HiddenMarkovModels.ObservationProcess

HiddenMarkovModels.StandardObservationProcess

HiddenMarkovModels.StandardStateProcess

HiddenMarkovModels.StateProcess

Base.length

Base.rand

DensityInterface.logdensityof

HiddenMarkovModels.baum_welch

HiddenMarkovModels.fit_element_from_sequence!

HiddenMarkovModels.forward_backward

HiddenMarkovModels.initial_distribution

HiddenMarkovModels.obs_distribution

HiddenMarkovModels.transition_matrix

HiddenMarkovModels.viterbi

StatsAPI.fit!