API reference

HiddenMarkovModels.HiddenMarkovModels

HiddenMarkovModels.HMM

HiddenMarkovModels.HiddenMarkovModel

HiddenMarkovModels.HiddenMarkovModel

HiddenMarkovModels.ObservationProcess

HiddenMarkovModels.StandardObservationProcess

HiddenMarkovModels.StandardStateProcess

HiddenMarkovModels.StateProcess

Base.rand

Base.rand

DensityInterface.logdensityof

HiddenMarkovModels.baum_welch

HiddenMarkovModels.fit_from_sequence

HiddenMarkovModels.forward_backward

HiddenMarkovModels.viterbi

HiddenMarkovModels.HiddenMarkovModels — Module

HiddenMarkovModels

A Julia package for HMM modeling, simulation, inference and learning.

HiddenMarkovModels.HMM — Type

HMM

Alias for the struct HiddenMarkovModel.

HiddenMarkovModels.HiddenMarkovModel — Type

HiddenMarkovModel{SP<:StateProcess,OP<:ObservationProcess}

Combination of a state and an observation process, amenable to simulation, inference and learning.

Fields

state_process::SP

obs_process::OP

HiddenMarkovModels.HiddenMarkovModel — Method

HMM(p, A, dists)

Construct an HMM from a vector of initial probabilities, a matrix of transition probabilities and a vector of observation distributions.

Same constructor as in HMMBase.jl.

HiddenMarkovModels.ObservationProcess — Type

ObservationProcess

Abstract type for the observation part of an HMM.

Required methods

Base.length(op)

distribution(op, i)

Optional methods

reestimate!(op, obs_seq, γ)

HiddenMarkovModels.StandardObservationProcess — Type

StandardObservationProcess{D} <: ObservationProcess

Fields

distributions::AbstractVector{D}: one distribution per state

HiddenMarkovModels.StandardStateProcess — Type

StandardStateProcess <: StateProcess

Fields

p::AbstractVector: initial distribution

A::AbstractMatrix: transition matrix

HiddenMarkovModels.StateProcess — Type

StateProcess

Abstract type for the state part of an HMM.

Required methods

Base.length(sp)

initial_distribution(sp)

transition_matrix(sp)

Optional methods

reestimate!(sp, p_count, A_count)

Base.rand — Method

rand(hmm, T)

Simulate an HMM for T time steps.

Base.rand — Method

rand(rng, hmm, T)

Simulate an HMM for T time steps with a specified rng.

DensityInterface.logdensityof — Method

DensityInterface.logdensityof(hmm, obs_seq)

Apply the forward algorithm to compute the loglikelihood of a sequence of observations.

HiddenMarkovModels.baum_welch — Method

baum_welch(hmm_init, obs_seqs; max_iterations, rtol)

Apply the Baum-Welch algorithm to estimate the parameters of an HMM on multiple observation sequences, and return a tuple (hmm, logL_evolution).

HiddenMarkovModels.fit_from_sequence — Method

fit_from_sequence(::Type{D}, x, w)

Fit a distribution of type D based on a single sequence of observations x associated with a single sequence of weights w.

Default to StatsAPI.fit, with a special case for Distributions.jl and vectors of vectors (because this implementation of fit accepts matrices instead). Users are free to override this default for concrete distributions.

HiddenMarkovModels.forward_backward — Method

forward_backward(hmm, obs_seq)

Apply the forward-backward algorithm to estimate the posterior state marginals of an HMM.

HiddenMarkovModels.viterbi — Method

viterbi(hmm, obs_seq)

Apply the Viterbi algorithm to compute the most likely sequence of states of an HMM.